// src/models/Category.ts
export enum Category {
  PRODUCE = "Produce",
  DAIRY_EGGS = "Dairy & Eggs",
  MEAT = "Meat",
  PREPARED_FOODS = "Prepared Foods",
  PANTRY_ESSENTIALS = "Pantry Essentials",
  BREADS_ROLLS_BAKERY = "Breads, Rolls & Bakery",
  DESSERTS = "Desserts",
  BODY_CARE = "Body Care",
  SUPPLEMENTS = "Supplements",
  FROZEN_FOODS = "Frozen Foods",
  SNACKS_CHIPS_SALSAS_DIPS = "Snacks, Chips, Salsas & Dips",
  SEAFOOD = "Seafood",
  BEVERAGES = "Beverages",
  WINE_BEER_SPIRITS = "Wine, Beer & Spirits",
  BEAUTY = "Beauty",
  FLORAL = "Floral",
  HOUSEHOLD = "Household",
  BABY_CHILD = "Baby & Child",
  LIFESTYLE = "Lifestyle",
  PET = "Pet",
}